Move get database connection a separate file
