+++ get_database_connection.py
@@ -1,0 +1,13 @@
+import os
+import redis
+
+_database = None
+
+def get_database_connection():
+    global _database
+    if _database is None:
+        database_password = os.getenv('DATABASE_PASSWORD')
+        database_host = os.getenv('DATABASE_HOST')
+        database_port = os.getenv('DATABASE_PORT')
+        _database = redis.Redis(host=database_host, port=database_port, password=database_password)
+    return _database
